% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastkqr-methods.R
\name{objective.fastkqr}
\alias{objective.fastkqr}
\title{Return the objective values for a \code{fastkqr} object.}
\usage{
objective.fastkqr(object, x, y, s = NULL)
}
\arguments{
\item{object}{A fitted \code{fastkqr} object.}

\item{x}{The predictor matrix, i.e., the \code{x} matrix used when fitting the \code{fastkqr} object.}

\item{y}{Response variable. The length is \eqn{n}.}

\item{s}{Value(s) of the penalty parameter `lambda` at which
predictions are required. Default is the entire sequence used to create the
model.}

\item{...}{Not used.}
}
\value{
Returns the objective values.
}
\description{
Return the objective values for a \code{fastkqr} object.
}
\examples{
library(MASS)
data(GAGurine)
x <- as.matrix(GAGurine$Age)
y <- GAGurine$GAG
lambda <- 10^(seq(1, -4, length.out=10))
fit <- fastkqr(x, y, lambda=lambda, tau=0.1)
objective.fastkqr(fit, x, y, lambda[1])
}
