% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastnckqr-methods.R
\name{objective.fastnckqr}
\alias{objective.fastnckqr}
\title{Calculate the objective values for a \code{fastnckqr} object.}
\usage{
objective.fastnckqr(object, x, y, s2, s1 = NULL)
}
\arguments{
\item{object}{A fitted \code{fastnckqr} object.}

\item{x}{The predictor matrix, i.e., the \code{x} matrix used when fitting the \code{fastnckqr} object.}

\item{y}{Response variable. The length is \eqn{n}.}

\item{s2}{Value of the penalty parameter `lambda2` at which
predictions are required.}

\item{s1}{Value(s) of the penalty parameter `lambda1` at which
predictions are required. Default is the entire sequence used to create the
model.}

\item{...}{Not used.}
}
\value{
Returns the objective values for the non-crossing kernel quantile regression model.
}
\description{
Calculate the objective values for a \code{fastnckqr} object.
}
\examples{
library(MASS)
data(GAGurine)
x <- as.matrix(GAGurine$Age)
y <- GAGurine$GAG
l2 <- 1e-4
ttau <- c(0.1, 0.3, 0.5, 0.7, 0.9)
l1_list <- 10^seq(-8, 2, length.out=10)
fit <- fastnckqr(x,y, lambda1=l1_list, lambda2=l2, tau=ttau)
objective.fastnckqr(fit, x, y, l2)
}
\keyword{kernel}
\keyword{regression}
